+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GroupsDetailPanel: implement posts paging [#572](https://github.com/rokwire/illinois-app/issues/572).
 - Show GroupPost reply thread [#581](https://github.com/rokwire/illinois-app/issues/581).
 - Show group post dates as time interval since now [#580](https://github.com/rokwire/illinois-app/issues/580).
+- Reply widget clean up [#584](https://github.com/rokwire/illinois-app/issues/584).
 
 ## [2.6.8] - 2021-07-22
 ### Changed